Extract Prompt Studio extend layout helpers
